Renamed picture with one more representative name
